Доработка домашней работы
