+++ main.py
@@ -1,7 +1,8 @@
 from typing import Any, Dict, List
 
 from src.processing import filter_by_state, sort_by_date
-from src.widget.mask_account_card import get_date, get_mask_account, get_mask_card_number
+from src.masks import get_mask_card_number
+from src.widget import get_date, get_mask_account
 
 
 def main() -> None:
